updated Makefile based on Intel Sample Code, fixed size_t to uint32_T cast warning

+++ sgx/trusted/include/kmyth_enclave_trusted.h
@@ -18,6 +18,7 @@ extern "C"
 
 #include "sgx_urts.h"
 
+#include <stdlib.h>
 #include <stdint.h>
 #include <stddef.h>
 #include <stdbool.h>
@@ -27,14 +28,14 @@ extern "C"
   typedef struct unseal_data_s
   {
     uint64_t handle;
-    size_t data_size;
+    uint32_t data_size;
     uint8_t *data;
     struct unseal_data_s *next;
   } unseal_data_t;
 
   extern unseal_data_t *kmyth_unsealed_data_table;
 
-  size_t retrieve_from_unseal_table(uint64_t handle, uint8_t ** buf);
+  uint32_t retrieve_from_unseal_table(uint64_t handle, uint8_t ** buf);
 
   bool insert_into_unseal_table(uint8_t * data,
                                 uint32_t data_size,
